requirements.txt
sentence-transformers
faiss-cpu
pypdf
numpy
torch
pytesseract
pdf2image